refactor(player): abstract player service
