+++ libs/feature/player/browser/src/components/player-exercise/player-exercise.component.html
@@ -72,7 +72,7 @@
                 <mat-panel-title>Solution</mat-panel-title>
               </mat-expansion-panel-header>
               <div class="exercise-form">
-                <nge-markdown [data]="player?.solution" />
+                <nge-markdown [data]="player!.solution!" />
               </div>
             </mat-expansion-panel>
           </mat-accordion>
