+++ libs/feature/player/common/src/index.ts
@@ -1,5 +1,8 @@
 export * from './lib/player-answer.model'
 export * from './lib/player-guards.model'
+export * from './lib/player-manager.model'
 export * from './lib/player-navigation.model'
 export * from './lib/player-renderer.model'
 export * from './lib/player.model'
+export * from './lib/sandbox-manager.model'
+export * from './lib/sandbox.model'
